requirements.txt
torch
transformers
datasets
Pillow
accelerate
bitsandbytes
sentencepiece
fastapi
uvicorn[standard]
python-multipart